backend/requirements.txt
fastapi
uvicorn
networkx
# optional
openai
# tree-sitter python bindings
tree_sitter
google-generativeai
python-dotenv
matplotlib
